# openapi/openapi.yml
components:
  schemas:
    Aggregate:
      description: Enumeration of aggregate values
      enum:
      - ok
      - test
      - staging
      - 'no'
      title: Aggregate
...
    DatabaseCreate:
      description: 'Model for creating new LinksResources representing `/databases`
        resources in the

        MongoDB.


        Required fields:


        - `name`

        - `base_url`


        Original required fields for a

        [`LinksResourceAttributes`](https://www.optimade.org/optimade-python-tools/api_reference/models/links/#optimade.models.links.LinksResourceAttributes)

        model:


        - `name`

        - `description`

        - `link_type`'
      properties:
        aggregate:
          allOf:
          - $ref: '#/components/schemas/Aggregate'
          default: ok
          description: 'A string indicating whether a client that is following links
            to aggregate results from different OPTIMADE implementations should follow
            this link or not.

            This flag SHOULD NOT be indicated for links where `link_type` is not `child`.


            If not specified, clients MAY assume that the value is `ok`.

            If specified, and the value is anything different than `ok`, the client
            MUST assume that the server is suggesting not to follow the link during
            aggregation by default (also if the value is not among the known ones,
            in case a future specification adds new accepted values).


            Specific values indicate the reason why the server is providing the suggestion.

            A client MAY follow the link anyway if it has reason to do so (e.g., if
            the client is looking for all test databases, it MAY follow the links
            marked with `aggregate`=`test`).


            If specified, it MUST be one of the values listed in section Link Aggregate
            Options.'
          title: Aggregate
        base_url:
          anyOf:
          - format: uri
            maxLength: 65536
            minLength: 1
            type: string
          - $ref: '#/components/schemas/Link'
          title: Base Url
        description:
          title: Description
          type: string
        homepage:
          anyOf:
          - format: uri
            maxLength: 65536
            minLength: 1
            type: string
          - $ref: '#/components/schemas/Link'
          description: JSON API links object, pointing to a homepage URL for this
            implementation.
          title: Homepage
        id:
          title: Id
          type: string
        immutable_id:
          description: "The entry's immutable ID (e.g., an UUID). This is important\
            \ for databases having preferred IDs that point to \"the latest version\"\
            \ of a record, but still offer access to older variants. This ID maps\
            \ to the version-specific record, in case it changes in the future.\n\n\
            - **Type**: string.\n\n- **Requirements/Conventions**:\n    - **Support**:\
            \ OPTIONAL support in implementations, i.e., MAY be `null`.\n    - **Query**:\
            \ MUST be a queryable property with support for all mandatory filter features.\n\
            \n- **Examples**:\n    - `\"8bd3e750-b477-41a0-9b11-3a799f21b44f\"`\n\
            \    - `\"fjeiwoj,54;@=%<>#32\"` (Strings that are not URL-safe are allowed.)"
          title: Immutable Id
          type: string
        last_modified:
          format: date-time
          title: Last Modified
          type: string
        link_type:
          allOf:
          - $ref: '#/components/schemas/LinkType'
          description: 'The type of the linked relation.

            MUST be one of these values: ''child'', ''root'', ''external'', ''providers''.'
          title: Link Type
        name:
          description: Human-readable name for the OPTIMADE API implementation, e.g.,
            for use in clients to show the name to the end-user.
          title: Name
          type: string
        no_aggregate_reason:
          description: 'An OPTIONAL human-readable string indicating the reason for
            suggesting not to aggregate results following the link.

            It SHOULD NOT be present if `aggregate`=`ok`.'
          title: No Aggregate Reason
          type: string
      required:
      - name
      - base_url
      title: DatabaseCreate
      type: object
    DatabasesResponse:
      description: "Successful response for `GET /databases`\n\nThis model is essentially\
        \ equal to\n[`LinksResponse`](https://www.optimade.org/optimade-python-tools/api_reference/models/responses/#optimade.models.responses.LinksResponse)\n\
        with the exception of the `data\xB4 field's description."
      properties:
        data:
          description: 'List of unique OPTIMADE links resource objects.

            These links resource objects represents OPTIMADE databases that can be
            used for queries in gateways.'
          items:
            $ref: '#/components/schemas/LinksResource'
          title: Data
          type: array
          uniqueItems: true
        errors:
          description: A list of unique errors
          items:
            $ref: '#/components/schemas/Error'
          title: Errors
          type: array
          uniqueItems: true
        included:
          anyOf:
          - items:
              $ref: '#/components/schemas/optimade__models__entries__EntryResource'
            type: array
          - items:
              type: object
            type: array
          title: Included
          uniqueItems: true
        jsonapi:
          allOf:
          - $ref: '#/components/schemas/JsonApi'
          description: Information about the JSON API used
          title: Jsonapi
        links:
          allOf:
          - $ref: '#/components/schemas/ToplevelLinks'
          description: Links associated with the primary data or errors
          title: Links
        meta:
          allOf:
          - $ref: '#/components/schemas/ResponseMeta'
          description: A meta object containing non-standard information
          title: Meta
      required:
      - data
      - meta
      title: DatabasesResponse
      type: object
    DatabasesResponseSingle:
      description: Successful response for `POST /databases` and `GET /databases/{database_id}`
      properties:
        data:
          allOf:
          - $ref: '#/components/schemas/LinksResource'
          description: 'A unique OPTIMADE links resource object.

            The OPTIMADE links resource object has just been created or found according
            to the specific query parameter(s) or URL id.

            It represents an OPTIMADE database that can be used for queries in gateways.'
          title: Data
        errors:
          description: A list of unique errors
          items:
            $ref: '#/components/schemas/Error'
          title: Errors
          type: array
          uniqueItems: true
        included:
          anyOf:
          - items:
              $ref: '#/components/schemas/optimade__models__entries__EntryResource'
            type: array
          - items:
              type: object
            type: array
          title: Included
          uniqueItems: true
        jsonapi:
          allOf:
          - $ref: '#/components/schemas/JsonApi'
          description: Information about the JSON API used
          title: Jsonapi
        links:
          allOf:
          - $ref: '#/components/schemas/ToplevelLinks'
          description: Links associated with the primary data or errors
          title: Links
        meta:
          allOf:
          - $ref: '#/components/schemas/ResponseMeta'
          description: A meta object containing non-standard information
          title: Meta
      required:
      - data
      - meta
      title: DatabasesResponseSingle
      type: object
...
    Error:
      description: An error response
      properties:
        code:
          description: an application-specific error code, expressed as a string value.
          title: Code
          type: string
        detail:
          description: A human-readable explanation specific to this occurrence of
            the problem.
          title: Detail
          type: string
        id:
          description: A unique identifier for this particular occurrence of the problem.
          title: Id
          type: string
        links:
          allOf:
          - $ref: '#/components/schemas/ErrorLinks'
          description: A links object storing about
          title: Links
        meta:
          allOf:
          - $ref: '#/components/schemas/Meta'
          description: a meta object containing non-standard meta-information about
            the error.
          title: Meta
        source:
          allOf:
          - $ref: '#/components/schemas/ErrorSource'
          description: An object containing references to the source of the error
          title: Source
        status:
          description: the HTTP status code applicable to this problem, expressed
            as a string value.
          title: Status
          type: string
        title:
          description: A short, human-readable summary of the problem. It **SHOULD
            NOT** change from occurrence to occurrence of the problem, except for
            purposes of localization.
          title: Title
          type: string
      title: Error
      type: object
...
    ErrorResponse:
      description: errors MUST be present and data MUST be skipped
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/Resource'
          - items:
              $ref: '#/components/schemas/Resource'
            type: array
          description: Outputted Data
          title: Data
          uniqueItems: true
        errors:
          description: A list of OPTIMADE-specific JSON API error objects, where the
            field detail MUST be present.
          items:
            $ref: '#/components/schemas/OptimadeError'
          title: Errors
          type: array
          uniqueItems: true
        included:
          description: A list of unique included resources
          items:
            $ref: '#/components/schemas/Resource'
          title: Included
          type: array
          uniqueItems: true
        jsonapi:
          allOf:
          - $ref: '#/components/schemas/JsonApi'
          description: Information about the JSON API used
          title: Jsonapi
        links:
          allOf:
          - $ref: '#/components/schemas/ToplevelLinks'
          description: Links associated with the primary data or errors
          title: Links
        meta:
          allOf:
          - $ref: '#/components/schemas/ResponseMeta'
          description: A meta object containing non-standard information.
          title: Meta
      required:
      - meta
      - errors
      title: ErrorResponse
      type: object
...
    JsonApi:
      description: An object describing the server's implementation
      properties:
        meta:
          allOf:
          - $ref: '#/components/schemas/Meta'
          description: Non-standard meta information
          title: Meta
        version:
          default: '1.0'
          description: Version of the json API used
          title: Version
          type: string
      title: JsonApi
      type: object
...
  securitySchemes:
    OAuth2AuthorizationCodeBearer:
      flows:
        authorizationCode:
          authorizationUrl: https://staging.the-marketplace.eu/oauth/oauth2/auth
          scopes: {}
          tokenUrl: https://staging.the-marketplace.eu/oauth/oauth2/token
      type: oauth2
info:
  description: A gateway server to query multiple OPTIMADE databases.
  title: OPTIMADE Gateway
  version: 0.2.2
  x-application-id: ''
  x-application-name: OPTIMADE Gateway
  x-external-hostname: http://localhost:8000
  x-products:
  - name: OPTIMADE Gateway
    productId: 22e6e756-7179-47c9-8f37-1ee06c7b38ae
openapi: 3.0.2
paths:
...
  /databases:
    get:
      description: '`GET /databases`


        Return overview of all (active) databases.'
      operationId: get_databases_databases_get
      parameters:
      - description: A filter string, in the format described in section API Filtering
          Format Specification of the specification.
        in: query
        name: filter
        required: false
        schema:
          default: ''
          description: A filter string, in the format described in section API Filtering
            Format Specification of the specification.
          title: Filter
          type: string
      - description: 'The output format requested (see section Response Format).

          Defaults to the format string ''json'', which specifies the standard output
          format described in this specification.

          Example: `http://example.com/v1/structures?response_format=xml`'
        in: query
        name: response_format
        required: false
        schema:
          default: json
          description: 'The output format requested (see section Response Format).

            Defaults to the format string ''json'', which specifies the standard output
            format described in this specification.

            Example: `http://example.com/v1/structures?response_format=xml`'
          title: Response Format
          type: string
      - description: 'An email address of the user making the request.

          The email SHOULD be that of a person and not an automatic system.

          Example: `http://example.com/v1/structures?email_address=user@example.com`'
        in: query
        name: email_address
        required: false
        schema:
          default: ''
          description: 'An email address of the user making the request.

            The email SHOULD be that of a person and not an automatic system.

            Example: `http://example.com/v1/structures?email_address=user@example.com`'
          format: email
          title: Email Address
          type: string
      - description: 'A comma-delimited set of fields to be provided in the output.

          If provided, these fields MUST be returned along with the REQUIRED fields.

          Other OPTIONAL fields MUST NOT be returned when this parameter is present.

          Example: `http://example.com/v1/structures?response_fields=last_modified,nsites`'
        in: query
        name: response_fields
        required: false
        schema:
          default: ''
          description: 'A comma-delimited set of fields to be provided in the output.

            If provided, these fields MUST be returned along with the REQUIRED fields.

            Other OPTIONAL fields MUST NOT be returned when this parameter is present.

            Example: `http://example.com/v1/structures?response_fields=last_modified,nsites`'
          pattern: ([a-z_][a-z_0-9]*(,[a-z_][a-z_0-9]*)*)?
          title: Response Fields
          type: string
      - description: 'If supporting sortable queries, an implementation MUST use the
          `sort` query parameter with format as specified by [JSON API 1.0](https://jsonapi.org/format/1.0/#fetching-sorting).


          An implementation MAY support multiple sort fields for a single query.

          If it does, it again MUST conform to the JSON API 1.0 specification.


          If an implementation supports sorting for an entry listing endpoint, then
          the `/info/<entries>` endpoint MUST include, for each field name `<fieldname>`
          in its `data.properties.<fieldname>` response value that can be used for
          sorting, the key `sortable` with value `true`.

          If a field name under an entry listing endpoint supporting sorting cannot
          be used for sorting, the server MUST either leave out the `sortable` key
          or set it equal to `false` for the specific field name.

          The set of field names, with `sortable` equal to `true` are allowed to be
          used in the "sort fields" list according to its definition in the JSON API
          1.0 specification.

          The field `sortable` is in addition to each property description and other
          OPTIONAL fields.

          An example is shown in the section Entry Listing Info Endpoints.'
        in: query
        name: sort
        required: false
        schema:
          default: ''
          description: 'If supporting sortable queries, an implementation MUST use
            the `sort` query parameter with format as specified by [JSON API 1.0](https://jsonapi.org/format/1.0/#fetching-sorting).


            An implementation MAY support multiple sort fields for a single query.

            If it does, it again MUST conform to the JSON API 1.0 specification.


            If an implementation supports sorting for an entry listing endpoint, then
            the `/info/<entries>` endpoint MUST include, for each field name `<fieldname>`
            in its `data.properties.<fieldname>` response value that can be used for
            sorting, the key `sortable` with value `true`.

            If a field name under an entry listing endpoint supporting sorting cannot
            be used for sorting, the server MUST either leave out the `sortable` key
            or set it equal to `false` for the specific field name.

            The set of field names, with `sortable` equal to `true` are allowed to
            be used in the "sort fields" list according to its definition in the JSON
            API 1.0 specification.

            The field `sortable` is in addition to each property description and other
            OPTIONAL fields.

            An example is shown in the section Entry Listing Info Endpoints.'
          pattern: ([a-z_][a-z_0-9]*(,[a-z_][a-z_0-9]*)*)?
          title: Sort
          type: string
      - description: 'Sets a numerical limit on the number of entries returned.

          See [JSON API 1.0](https://jsonapi.org/format/1.0/#fetching-pagination).

          The API implementation MUST return no more than the number specified.

          It MAY return fewer.

          The database MAY have a maximum limit and not accept larger numbers (in
          which case an error code -- 403 Forbidden -- MUST be returned).

          The default limit value is up to the API implementation to decide.

          Example: `http://example.com/optimade/v1/structures?page_limit=100`'
        in: query
        name: page_limit
        required: false
        schema:
          default: 20
          description: 'Sets a numerical limit on the number of entries returned.

            See [JSON API 1.0](https://jsonapi.org/format/1.0/#fetching-pagination).

            The API implementation MUST return no more than the number specified.

            It MAY return fewer.

            The database MAY have a maximum limit and not accept larger numbers (in
            which case an error code -- 403 Forbidden -- MUST be returned).

            The default limit value is up to the API implementation to decide.

            Example: `http://example.com/optimade/v1/structures?page_limit=100`'
          minimum: 0.0
          title: Page Limit
          type: integer
      - description: 'RECOMMENDED for use with _offset-based_ pagination: using `page_offset`
          and `page_limit` is RECOMMENDED.

          Example: Skip 50 structures and fetch up to 100: `/structures?page_offset=50&page_limit=100`.'
        in: query
        name: page_offset
        required: false
        schema:
          default: 0
          description: 'RECOMMENDED for use with _offset-based_ pagination: using
            `page_offset` and `page_limit` is RECOMMENDED.

            Example: Skip 50 structures and fetch up to 100: `/structures?page_offset=50&page_limit=100`.'
          minimum: 0.0
          title: Page Offset
          type: integer
      - description: 'RECOMMENDED for use with _page-based_ pagination: using `page_number`
          and `page_limit` is RECOMMENDED.

          It is RECOMMENDED that the first page has number 1, i.e., that `page_number`
          is 1-based.

          Example: Fetch page 2 of up to 50 structures per page: `/structures?page_number=2&page_limit=50`.'
        in: query
        name: page_number
        required: false
        schema:
          default: 0
          description: 'RECOMMENDED for use with _page-based_ pagination: using `page_number`
            and `page_limit` is RECOMMENDED.

            It is RECOMMENDED that the first page has number 1, i.e., that `page_number`
            is 1-based.

            Example: Fetch page 2 of up to 50 structures per page: `/structures?page_number=2&page_limit=50`.'
          minimum: 0.0
          title: Page Number
          type: integer
      - description: 'RECOMMENDED for use with _cursor-based_ pagination: using `page_cursor`
          and `page_limit` is RECOMMENDED.'
        in: query
        name: page_cursor
        required: false
        schema:
          default: 0
          description: 'RECOMMENDED for use with _cursor-based_ pagination: using
            `page_cursor` and `page_limit` is RECOMMENDED.'
          minimum: 0.0
          title: Page Cursor
          type: integer
      - description: 'RECOMMENDED for use with _value-based_ pagination: using `page_above`/`page_below`
          and `page_limit` is RECOMMENDED.

          Example: Fetch up to 100 structures above sort-field value 4000 (in this
          example, server chooses to fetch results sorted by increasing `id`, so `page_above`
          value refers to an `id` value): `/structures?page_above=4000&page_limit=100`.'
        in: query
        name: page_above
        required: false
        schema:
          default: 0
          description: 'RECOMMENDED for use with _value-based_ pagination: using `page_above`/`page_below`
            and `page_limit` is RECOMMENDED.

            Example: Fetch up to 100 structures above sort-field value 4000 (in this
            example, server chooses to fetch results sorted by increasing `id`, so
            `page_above` value refers to an `id` value): `/structures?page_above=4000&page_limit=100`.'
          minimum: 0.0
          title: Page Above
          type: integer
      - description: 'RECOMMENDED for use with _value-based_ pagination: using `page_above`/`page_below`
          and `page_limit` is RECOMMENDED.'
        in: query
        name: page_below
        required: false
        schema:
          default: 0
          description: 'RECOMMENDED for use with _value-based_ pagination: using `page_above`/`page_below`
            and `page_limit` is RECOMMENDED.'
          minimum: 0.0
          title: Page Below
          type: integer
      - description: 'A server MAY implement the JSON API concept of returning [compound
          documents](https://jsonapi.org/format/1.0/#document-compound-documents)
          by utilizing the `include` query parameter as specified by [JSON API 1.0](https://jsonapi.org/format/1.0/#fetching-includes).


          All related resource objects MUST be returned as part of an array value
          for the top-level `included` field, see the section JSON Response Schema:
          Common Fields.


          The value of `include` MUST be a comma-separated list of "relationship paths",
          as defined in the [JSON API](https://jsonapi.org/format/1.0/#fetching-includes).

          If relationship paths are not supported, or a server is unable to identify
          a relationship path a `400 Bad Request` response MUST be made.


          The **default value** for `include` is `references`.

          This means `references` entries MUST always be included under the top-level
          field `included` as default, since a server assumes if `include` is not
          specified by a client in the request, it is still specified as `include=references`.

          Note, if a client explicitly specifies `include` and leaves out `references`,
          `references` resource objects MUST NOT be included under the top-level field
          `included`, as per the definition of `included`, see section JSON Response
          Schema: Common Fields.


          > **Note**: A query with the parameter `include` set to the empty string
          means no related resource objects are to be returned under the top-level
          field `included`.'
        in: query
        name: include
        required: false
        schema:
          default: references
          description: 'A server MAY implement the JSON API concept of returning [compound
            documents](https://jsonapi.org/format/1.0/#document-compound-documents)
            by utilizing the `include` query parameter as specified by [JSON API 1.0](https://jsonapi.org/format/1.0/#fetching-includes).


            All related resource objects MUST be returned as part of an array value
            for the top-level `included` field, see the section JSON Response Schema:
            Common Fields.


            The value of `include` MUST be a comma-separated list of "relationship
            paths", as defined in the [JSON API](https://jsonapi.org/format/1.0/#fetching-includes).

            If relationship paths are not supported, or a server is unable to identify
            a relationship path a `400 Bad Request` response MUST be made.


            The **default value** for `include` is `references`.

            This means `references` entries MUST always be included under the top-level
            field `included` as default, since a server assumes if `include` is not
            specified by a client in the request, it is still specified as `include=references`.

            Note, if a client explicitly specifies `include` and leaves out `references`,
            `references` resource objects MUST NOT be included under the top-level
            field `included`, as per the definition of `included`, see section JSON
            Response Schema: Common Fields.


            > **Note**: A query with the parameter `include` set to the empty string
            means no related resource objects are to be returned under the top-level
            field `included`.'
          title: Include
          type: string
      - description: If the client provides the parameter, the value SHOULD have the
          format `vMAJOR` or `vMAJOR.MINOR`, where MAJOR is a major version and MINOR
          is a minor version of the API. For example, if a client appends `api_hint=v1.0`
          to the query string, the hint provided is for major version 1 and minor
          version 0.
        in: query
        name: api_hint
        required: false
        schema:
          default: ''
          description: If the client provides the parameter, the value SHOULD have
            the format `vMAJOR` or `vMAJOR.MINOR`, where MAJOR is a major version
            and MINOR is a minor version of the API. For example, if a client appends
            `api_hint=v1.0` to the query string, the hint provided is for major version
            1 and minor version 0.
          pattern: (v[0-9]+(\.[0-9]+)?)?
          title: Api Hint
          type: string
      responses:
        '200':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/DatabasesResponse'
          description: Successful Response
        '400':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '422':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '553':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Version Not Supported
      security:
      - OAuth2AuthorizationCodeBearer: []
      summary: Get Databases
      tags:
      - Databases
    post:
      description: '`POST /databases`


        Create/Register or return an existing

        [`LinksResource`](https://www.optimade.org/optimade-python-tools/api_reference/models/links/#optimade.models.links.LinksResource),

        representing a database resource object, according to `database`.'
      operationId: post_databases_databases_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatabaseCreate'
        required: true
      responses:
        '200':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/DatabasesResponseSingle'
          description: Successful Response
        '400':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '422':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '553':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Version Not Supported
      security:
      - OAuth2AuthorizationCodeBearer: []
      summary: Post Databases
      tags:
      - Databases
...
servers:
- url: http://localhost:8000